+++ lib/mixins/jumper.dart
@@ -40,10 +40,12 @@ mixin Jumper on Movement, BlockMovementCollision {
   @override
   void onBlockedMovement(
     PositionComponent other,
-    Direction direction,
+    CollisionData collisionData,
   ) {
-    super.onBlockedMovement(other, direction);
-    if (jumping && lastDirectionVertical.isDownSide && direction.isDownSide) {
+    super.onBlockedMovement(other, collisionData);
+    if (jumping &&
+        lastDirectionVertical.isDownSide &&
+        collisionData.direction.isDownSide) {
       _currentJumps = 0;
       jumping = false;
     }
@@ -52,7 +54,7 @@ mixin Jumper on Movement, BlockMovementCollision {
   @override
   void update(double dt) {
     super.update(dt);
-    if (!jumping && lastDisplacement.y > 1) {
+    if (!jumping && displacement.y.abs() > 0.5) {
       jumping = true;
     }
     _notifyJump();
@@ -72,6 +74,13 @@ mixin Jumper on Movement, BlockMovementCollision {
     if (newDirection != _lastDirectionJump) {
       _lastDirectionJump = newDirection;
       onJump(newDirection);
+    }
+  }
+
+  @override
+  void stopMove({bool forceIdle = false, bool isX = true, bool isY = true}) {
+    if (!jumping) {
+      super.stopMove(forceIdle: forceIdle, isX: isX, isY: isY);
     }
   }
 }